feat: basic UI refractoring and update UI and new component for StudentQuestion
